1.修改deamon进程使得子进程等待失败。 2.修改monitor进程监控方式。 3.修改system函数产生的信号中断recvfrom函数。 4.优化安装和卸载脚本。 5.增加停止和启动脚本。

+++ server/bin/uninstall.sh
@@ -4,6 +4,15 @@ install_dir="/usr/local"
 DEAMON_DIR="${install_dir}/hot_backup_server"
 proc_name="hot_backup_server"
 
-pgrep ${proc_name} | xargs -t -I{} kill -9 {}
+ps axf | grep ${proc_name} | grep -v grep | awk '{print $1}' | xargs -t -I{} kill -9 {}
+
+pids=`ps axf | grep ${proc_name} | grep -v grep | awk '{print $1}'`
+if [ "$pids" != "" ]
+then
+    echo "stop ${proc_name} failed!"
+    exit 1
+fi
 
 rm -rf ${DEAMON_DIR}
+echo "uninstall ${proc_name} success!"
+exit 0
